src/main/java/org/apache/directory/shared/ldap/schema/AbstractSchemaObject.java : Modified the toString() method for better lisibility in debug mode
src/main/java/org/apache/directory/shared/ldap/util/AttributeUtils.java :
Added a method containsValueCaseIgnore( Attribute attr, Object value ) which search for a value in
an attribute, case insensitive. This method should be removed, and we should use the static method
containsValue( Attribute attr, Object compared, AttributeType type ) instead. But in 1.0, there is
two problems : 
1) We don't always have access to the AttributeType when we need to call this method (because we don't
have access to the registries)
2) It throws a NamingException and the caller does not handle this exception.
So the added method can be used for Attribute which are not caseSensitive, like ObjectClass.



git-svn-id: https://svn.apache.org/repos/asf/directory/shared/branches/0.9.5@498449 13f79535-47bb-0310-9956-ffa450edef68