Fixed DIRSERVER-823.
The problem was hidden in the way we handled the filter tree. It was supposed that when a filter has been
completly parsed, and if it's parent TLV expected length was 0, then the parent filter was considered as completed too.
But the fact is that we may have had some cases (like with substrings) where the parent TLV has an expected Length 
equals to 0 but this parent TLV was itself included in another parent TLV, representing the parent filter.

So we unfolded two filters instead of only one, and associated the completed filter to the wrong parent.

Now, an TLV id has been created, associated to each TLV in a container, and the unfolding is done comparing 
those ids.

More doco about this complex algorithm to come on confluence ...


git-svn-id: https://svn.apache.org/repos/asf/directory/shared/branches/0.9.5@503333 13f79535-47bb-0310-9956-ffa450edef68