reapplying all the commits from rev. 1051869, 1051970, 1053085, 1054545, 1055045, 1055125 and 1054996 all these are missing
(probably due to a mistake while merging)
o Added a method and a test to get the closest parent in a DnNode which has an element from a starting point.
o The getParentWithElement should return the element, not the DN
o Modified the getParentWithElement() method to return the node contaning an element instead of just returning the element.
o added rename operation and a test
o added move operation and tests

and the latest change besides the above merge

o replaced the ConcurrentHashmap to HashMap (DIRSHARED-73)


git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@1124303 13f79535-47bb-0310-9956-ffa450edef68