Normalizers and SyntaxCheckers are also modified like Comparators to be extendable in OSGI, and work normal in nonOSGI environment.
shared-ldap-schema-data is cleaned of code duplication which is used to make it work on both environments.
Detection mechanism is used, and code duplicates are deleted. Only classLoad... family of functions in SchemaEntityFactory
are changed to detect whether or not they're in OSGI or not, and act according to that test.



git-svn-id: https://svn.apache.org/repos/asf/directory/shared/branches/shared-osgi@1188705 13f79535-47bb-0310-9956-ffa450edef68